+++ myconext-server/src/test/java/myconext/AbstractIntegrationTest.java
@@ -66,6 +66,7 @@ import static org.junit.Assert.assertTrue;
                 "sp_entity_id=https://engine.test.surfconext.nl/authentication/sp/metadata",
                 "sp_entity_metadata_url=https://engine.test.surfconext.nl/authentication/sp/metadata",
                 "spring.main.lazy-initialization=true",
+                "eduid_api.oidcng_introspection_uri=http://localhost:8098/introspect",
                 "cron.service-name-resolver-initial-delay-milliseconds=60000"
         })
 @ActiveProfiles({"test"})